+++ HISTORY.rst
@@ -4,7 +4,7 @@
 
 - initial support for Robotice Control(management Robotice through django admin)
 - added platform utils for work with specific platforms (BB, RPi, ..)
-- support for more db backend (MongoDB, ..)
+- support for more db backends (MongoDB, ..)
 
 0.1 (2014-08-25)
 ================